Move format_color to colors.rs including tests
